requirements.txt
boto3==1.9.21
botocore==1.12.21
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.35.0
django-bootstrap-form==3.4
django-ckeditor==5.5.0
django-debug-toolbar==1.9.1
django-js-asset==1.1.0
django-social-widgets==0.4.0
django-storages==1.7.1
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
docutils==0.14
idna==2.6
jmespath==0.9.3
oauthlib==2.0.7
Pillow==5.1.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.7.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.13
six==1.11.0
sqlparse==0.2.4
urllib3==1.22
